smart-life-cloud/backend/app/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlmodel==0.0.21
SQLAlchemy==2.0.31
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.8.2
pydantic-settings==2.4.0
python-multipart==0.0.9
